[feature] Adding clone of a cluster in a different namespace
The clone take into account additional configs for pgbackrest. If there
are configmaps or secrets, it dumps them into the new namespace.

In case of failure at any stage, created resources are deleted.

If the target namespace does not exist, the code creates it. Note that,
in case of later failure, the namespace is NOT deleted. It is the only
resource that is not deleted.