[improvements] Add 2 options : show yaml and override resources
The show-yaml option allows to display the definition of the yaml of the
clone on the stdout.
The override-configs option allows to override the configmaps or secrets
required to be created when generating the clone. If, for any reason,
one wants to override the objects, one can use this option and the
secrets and configmaps will be overrided.