[clone] WIP : handling PITR in clone process
First step is to refactor the show command to list the backups. This
command can then be used in the clone process : we are listing the
backups, transform them to an object and check if the pitr requested is
reachable with respect to the different backup schedules.
Added some tests to validate the basis of the process.
Still some work to make.