[clone] Add new parameter to set the clone name
If one wants to choose the clone name, one can use the new parameter
--with-name which allows the user to set the clone name. This may be
useful to do the following trick : clone the source cluster that could
be harmed, check the clone. If the clone is fine, drop the source
cluster and clone the cloned cluster with the name of the original
cluster. And you have restored your cluster. Credit to a friend that
required this option :-) This may save time in case of heavily damaged
cluster...