[to be squashed] Temporary commit to delete secrets in configuration
In pgbackrest.global.configuration , dropping secrets. In my case,
secrets are pointing at S3 data and, if present, will make pgbackrest
assume that there is an incomplete repo2 configuration and it will break
the process. It should be more precise and not drop secrets
automatically but only if it contains specific configuration.